Fix RocksDB DB::Open API signature mismatch in auth modules

+++ src/api/geo_index_hooks.cpp
@@ -70,7 +70,7 @@ static bool validateGeoJSONBasic(const json& geojson) {
 // Helper function to validate and sanitize coordinate pair
 [[maybe_unused]] static bool validateCoordinatePair(const json& coord, double& lon, double& lat) {
     try {
-        if (!coord.is_array() || coord.size() < 2) {
+        if (!coord.is_array() || static_cast<int>(coord.size()) < 2) {
             return false;
         }
         
@@ -121,7 +121,7 @@ void GeoIndexHooks::onEntityPut(
         
         if (likely_json) {
             try {
-                std::string blob_str(reinterpret_cast<const char*>(blob.data()), blob.size());
+                std::string blob_str(reinterpret_cast<const char*>(blob.data()),static_cast<int>(blob.size()));
                 j = nlohmann::json::parse(blob_str);
             } catch (const nlohmann::json::exception &) {
                 throw;
@@ -145,7 +145,7 @@ void GeoIndexHooks::onEntityPut(
         if (j.contains("geometry") && j["geometry"].is_string()) {
             std::string hex_ewkb = j["geometry"].get<std::string>();
             geom_blob.reserve(hex_ewkb.size() / 2);
-            for (size_t i = 0; i + 1 < hex_ewkb.size(); i += 2) {
+            for (size_t i = 0; i + 1 <static_cast<int>(hex_ewkb.size()); i += 2) {
                 uint8_t byte = static_cast<uint8_t>(std::stoi(hex_ewkb.substr(i, 2), nullptr, 16));
                 geom_blob.push_back(byte);
             }
@@ -288,7 +288,7 @@ bool GeoIndexHooks::onEntityPutAtomic(
 
     try {
         // Parse blob as JSON to extract geometry field
-        std::string blob_str(reinterpret_cast<const char*>(blob.data()), blob.size());
+        std::string blob_str(reinterpret_cast<const char*>(blob.data()),static_cast<int>(blob.size()));
         auto j = json::parse(blob_str);
 
         // Look for geometry field
@@ -298,7 +298,7 @@ bool GeoIndexHooks::onEntityPutAtomic(
         if (j.contains("geometry") && j["geometry"].is_string()) {
             std::string hex_ewkb = j["geometry"].get<std::string>();
             geom_blob.reserve(hex_ewkb.size() / 2);
-            for (size_t i = 0; i < hex_ewkb.size(); i += 2) {
+            for (size_t i = 0; i <static_cast<int>(hex_ewkb.size()); i += 2) {
                 uint8_t byte = static_cast<uint8_t>(
                     std::stoi(hex_ewkb.substr(i, 2), nullptr, 16)
                 );
@@ -400,7 +400,7 @@ bool GeoIndexHooks::onEntityDeleteAtomic(
 
     try {
         // Parse old blob to extract geometry and compute sidecar
-        std::string blob_str(reinterpret_cast<const char*>(old_blob.data()), old_blob.size());
+        std::string blob_str(reinterpret_cast<const char*>(old_blob.data()),static_cast<int>(old_blob.size()));
         auto j = json::parse(blob_str);
 
         // Look for geometry field
@@ -411,7 +411,7 @@ bool GeoIndexHooks::onEntityDeleteAtomic(
         if (j.contains("geometry") && j["geometry"].is_string()) {
             std::string hex_ewkb = j["geometry"].get<std::string>();
             geom_blob.reserve(hex_ewkb.size() / 2);
-            for (size_t i = 0; i < hex_ewkb.size(); i += 2) {
+            for (size_t i = 0; i <static_cast<int>(hex_ewkb.size()); i += 2) {
                 uint8_t byte = static_cast<uint8_t>(
                     std::stoi(hex_ewkb.substr(i, 2), nullptr, 16)
                 );
@@ -518,7 +518,7 @@ void GeoIndexHooks::onEntityDelete(
         if (j.contains("geometry") && j["geometry"].is_string()) {
             std::string hex_ewkb = j["geometry"].get<std::string>();
             geom_blob.reserve(hex_ewkb.size() / 2);
-            for (size_t i = 0; i < hex_ewkb.size(); i += 2) {
+            for (size_t i = 0; i <static_cast<int>(hex_ewkb.size()); i += 2) {
                 uint8_t byte = static_cast<uint8_t>(
                     std::stoi(hex_ewkb.substr(i, 2), nullptr, 16)
                 );
