+++ src/rpc_grpc/bidi_stream_adapter.h
@@ -113,7 +113,7 @@ public:
             std::unique_lock<std::mutex> lock(queue_mutex_);
             // Wait while queue is full and stream is still open
             queue_not_full_.wait(lock, [this] {
-                return finished_ || queue_.size() < max_queue_depth_;
+                return finished_ || static_cast<int>(queue_.size()) < max_queue_depth_;
             });
 
             if (finished_) {
@@ -163,7 +163,7 @@ public:
      */
     std::size_t queueDepth() const {
         std::lock_guard<std::mutex> lock(queue_mutex_);
-        return queue_.size();
+        return static_cast<int>(queue_.size());
     }
 
     /**
